# backend/src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: MERMAID API
  version: 1.0.0
  description: |
    MERMAID (Marine Safety and Market Information Dashboard) API.
    MVP scope includes authentication, advisories, marine conditions,
    vendor demand listings, fisherman marketplace, trip sessions, and catch logs.
servers:
  - url: http://localhost:8080/api
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /buyer/orders/{orderId}/review:
    post:
      tags: [Reviews]
      summary: Create a review for a completed order
      operationId: createOrderReview
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          description: Order not found or not yours
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Order not yet completed or already reviewed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Reviews]
      summary: Get current buyer's review for this order
      operationId: getOrderReview
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Review (if any)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          description: No review yet
    patch:
      tags: [Reviews]
      summary: Update buyer's review (within 7 days of creation)
      operationId: updateOrderReview
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      responses:
        '200':
          description: Updated review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '409':
          description: Review locked (older than 7 days)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /vendors/{vendorId}/storefront:
    get:
      tags: [Reviews]
      summary: Public vendor storefront (profile + open listings + recent reviews)
      operationId: getVendorStorefront
      security: []
      parameters:
        - in: path
          name: vendorId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Storefront
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorStorefront'
        '404':
          description: Vendor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /vendors/{vendorId}/reviews:
    get:
      tags: [Reviews]
      summary: Public list of reviews for a vendor
      operationId: getVendorReviews
      security: []
      parameters:
        - in: path
          name: vendorId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Paginated reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedReviews'
...
  /orders:
    post:
      tags: [Orders]
      summary: Create an order / make an offer on a catch alert (Vendor)
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Catch alert or demand listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/mine:
    get:
      tags: [Orders]
      summary: List orders where current user is buyer or seller
      operationId: listMyOrders
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, CONFIRMED, COMPLETED, CANCELLED, DISPUTED]
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /orders/{orderId}/confirm:
    put:
      tags: [Orders]
      summary: Confirm (accept) an order (Fisherman / seller)
      operationId: confirmOrder
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Order not in PENDING state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{orderId}/cancel:
    put:
      tags: [Orders]
      summary: Cancel an order (buyer or seller)
      operationId: cancelOrder
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{orderId}/handoff:
    post:
      tags: [Orders]
      summary: Create handoff confirmation record (either party)
      operationId: createHandoff
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HandoffCreateRequest'
      responses:
        '201':
          description: Handoff created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandoffConfirmation'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Order not in CONFIRMED state or handoff already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{orderId}/handoff/confirm-seller:
    put:
      tags: [Orders]
      summary: Seller (fisherman) confirms handoff weight
      operationId: confirmHandoffSeller
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Seller confirmation recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandoffConfirmation'
        '404':
          description: Order or handoff not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{orderId}/handoff/confirm-buyer:
    put:
      tags: [Orders]
      summary: Buyer (vendor) confirms handoff weight
      operationId: confirmHandoffBuyer
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Buyer confirmation recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandoffConfirmation'
        '404':
          description: Order or handoff not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{orderId}/payment:
    post:
      tags: [Orders]
      summary: Record payment (Vendor / buyer)
      operationId: recordPayment
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreateRequest'
      responses:
        '201':
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRecord'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Payment already exists or handoff not confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{orderId}/payment/confirm:
    put:
      tags: [Orders]
      summary: Confirm payment receipt (Fisherman / seller)
      operationId: confirmPayment
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRecord'
        '404':
          description: Order or payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /orders/{orderId}/timeline:
    get:
      tags: [Orders]
      summary: Get order status timeline (any participant)
      operationId: getOrderTimeline
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Timeline events for the order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderStatusEvent'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /webhooks/paymongo:
    post:
      tags: [Webhooks]
      summary: PayMongo webhook handler
      operationId: handlePaymongoWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderId:
      in: path
      name: orderId
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    DemandListingStatus:
      type: string
      enum: [OPEN, CLOSED]
...
    FishSpecies:
      type: object
      required: [id, commonName, active]
      properties:
        id:
          type: integer
          format: int64
        commonName:
          type: string
        scientificName:
          type: string
          nullable: true
        active:
          type: boolean
...
    MarketLocation:
      type: object
      required: [id, name, municipality, active]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        municipality:
          type: string
        province:
          type: string
          nullable: true
        active:
          type: boolean
        lat:
          type: number
          format: double
          nullable: true
        lng:
          type: number
          format: double
          nullable: true
...
    DemandListing:
      type: object
      required:
        [id, vendorId, fishSpecies, marketLocation, quantityKg, offerPricePerKg, status, postedAt]
      properties:
        id:
          type: integer
          format: int64
        vendorId:
          type: integer
          format: int64
        vendorName:
          type: string
          nullable: true
        fishSpecies:
          $ref: '#/components/schemas/FishSpecies'
        marketLocation:
          $ref: '#/components/schemas/MarketLocation'
        quantityKg:
          type: number
          format: double
        offerPricePerKg:
          type: number
          format: double
        notes:
          type: string
          nullable: true
        neededBy:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/DemandListingStatus'
        postedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        distanceKm:
          type: number
          format: double
          nullable: true
          description: Haversine distance from buyer's lat/lng (when supplied) to listing market location
...
    BuyerVendorProfile:
      type: object
      required: [id, fullName]
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        avgRating:
          type: number
          format: double
          nullable: true
          description: Average review rating (null until reviews are implemented)
        reviewCount:
          type: integer
          format: int64
          nullable: true
        joinedDate:
          type: string
          format: date-time
          nullable: true
        totalCompletedTrades:
          type: integer
          format: int64
          nullable: true
    VendorStorefront:
      type: object
      required: [vendor, listings, recentReviews]
      properties:
        vendor:
          $ref: '#/components/schemas/BuyerVendorProfile'
        listings:
          type: array
          items:
            $ref: '#/components/schemas/DemandListing'
        recentReviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
    Review:
      type: object
      required: [id, orderId, reviewerId, vendorId, rating, createdAt]
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        reviewerId:
          type: integer
          format: int64
        reviewerName:
          type: string
          nullable: true
        vendorId:
          type: integer
          format: int64
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        photoUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        editable:
          type: boolean
          description: True when within the 7-day edit window for this reviewer
    CreateReviewRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          maxLength: 2000
        photoUrls:
          type: array
          items:
            type: string
    UpdateReviewRequest:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          maxLength: 2000
        photoUrls:
          type: array
          items:
            type: string
    PagedReviews:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        avgRating:
          type: number
          format: double
          nullable: true
...
    UserRef:
      type: object
      required: [id, fullName]
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
...
    OrderCreateRequest:
      type: object
      required: [speciesId, agreedPricePerKg]
      properties:
        catchAlertId:
          type: integer
          format: int64
          nullable: true
        demandListingId:
          type: integer
          format: int64
          nullable: true
        speciesId:
          type: integer
          format: int64
        agreedPricePerKg:
          type: number
          format: double
          minimum: 0
        orderedQtyEstimate:
          type: string
          maxLength: 100
          nullable: true
        orderedQtyKg:
          type: number
          format: double
          minimum: 0.1
          nullable: true
        dispatchMode:
          type: string
          enum: [PICKUP, DELIVERY]
          nullable: true
        notes:
          type: string
          maxLength: 500
          nullable: true
    Order:
      type: object
      required: [id, buyer, seller, species, agreedPricePerKg, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
        buyer:
          $ref: '#/components/schemas/UserRef'
        seller:
          $ref: '#/components/schemas/UserRef'
        species:
          $ref: '#/components/schemas/FishSpecies'
        catchAlertId:
          type: integer
          format: int64
          nullable: true
        demandListingId:
          type: integer
          format: int64
          nullable: true
        agreedPricePerKg:
          type: number
          format: double
        orderedQtyEstimate:
          type: string
          nullable: true
        orderedQtyKg:
          type: number
          format: double
          nullable: true
        dispatchMode:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, CONFIRMED, COMPLETED, CANCELLED, DISPUTED]
        notes:
          type: string
          nullable: true
        deliveryAddress:
          type: string
          nullable: true
        handoff:
          $ref: '#/components/schemas/HandoffConfirmation'
          nullable: true
        payment:
          $ref: '#/components/schemas/PaymentRecord'
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    HandoffCreateRequest:
      type: object
      required: [actualQtyKg, finalPricePerKg]
      properties:
        actualQtyKg:
          type: number
          format: double
          minimum: 0.1
        finalPricePerKg:
          type: number
          format: double
          minimum: 0
    HandoffConfirmation:
      type: object
      required: [id, orderId, actualQtyKg, finalPricePerKg, totalAmount, confirmedByBuyer, confirmedBySeller, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        actualQtyKg:
          type: number
          format: double
        finalPricePerKg:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        confirmedByBuyer:
          type: boolean
        confirmedBySeller:
          type: boolean
        disputeReason:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, CONFIRMED, DISPUTED]
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    PaymentCreateRequest:
      type: object
      required: [amount, method]
      properties:
        amount:
          type: number
          format: double
          minimum: 0
        method:
          type: string
          enum: [CASH, GCASH, MAYA, COD, CREDIT, BANK_TRANSFER]
        proofReference:
          type: string
          maxLength: 200
          nullable: true
    PaymentRecord:
      type: object
      required: [id, orderId, amount, method, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        handoffId:
          type: integer
          format: int64
          nullable: true
        amount:
          type: number
          format: double
        method:
          type: string
        proofReference:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, CONFIRMED, DISPUTED]
        paidAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        code:
          type: string
          nullable: true
        fieldErrors:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...
    OrderStatusEvent:
      type: object
      required: [id, orderId, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        status:
          type: string
          description: The status the order transitioned to
        actorId:
          type: integer
          format: int64
          nullable: true
        actorName:
          type: string
          nullable: true
          description: Resolved name of the actor who triggered the transition
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...